[test] move prod debug unlock test to its own file
Also:
1. ensure the manuf debug unlock test runs in CI, and
2. rev the caliptra-sw repo version.
